+++ core/src/main/scala/caliban/parsing/adt/Value.scala
@@ -4,8 +4,8 @@ sealed trait Value
 
 object Value {
   case object NullValue                              extends Value
-  case class IntValue(value: Int)                    extends Value
-  case class FloatValue(value: Float)                extends Value
+  case class IntValue(value: Long)                   extends Value
+  case class FloatValue(value: Double)               extends Value
   case class StringValue(value: String)              extends Value
   case class BooleanValue(value: Boolean)            extends Value
   case class EnumValue(value: String)                extends Value
